fix: default imagePullPolicy for Plugin Deployment to Always (#178)
## Description
Updates the template for the Plugin Registry deployment to allow user to
configure the imagePullPolicy, and sets the policy to `Always` by
default. Additionally allows users to opt out of deploying the plugin
registry in the event they are using a fleeting plugin that is not
included in the plugin registry image.

## Related Issue

Fixes #177 

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/uds-package-gitlab-runner/blob/main/CONTRIBUTING.md#developer-workflow)
followed